+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
+## [0.15.0] - 2026-04-22
+
+### Added
+- contract names
+- suggests pause-capable contracts
+
 ## [Unreleased]
+
+### Added
+- suggests pause-capable contracts
 
 ## [0.14.0] - 2026-04-22
 ### Added